pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = ""
authors = [
    {name = "Alissa",email = "alissajouljian@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain (>=0.3.27,<0.4.0)",
    "langchain-google-genai (>=2.1.12,<3.0.0)",
    "langchain-community (>=0.3.30,<0.4.0)",
    "google-search-results (>=2.4.2,<3.0.0)",
    "streamlit (>=1.50.0,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "chromadb (>=1.1.0,<2.0.0)",
    "openai (>=1.109.1,<2.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "langchain-chroma (>=0.2.6,<0.3.0)",
    "langchain-openai (>=0.3.33,<0.4.0)",
    "numpy (<2)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"

